backend/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
groq
httpx
sentence-transformers
pinecone
PyMuPDF
tqdm
Pillow
psycopg2-binary
gunicorn
whitenoise
python-dotenv
